+++ resources/js/components/modals/AddNewMenuItemModal.vue
@@ -1,23 +1,23 @@
 <template>
-  <modal :show="showModal" :align="'flex justify-end'" class="add-new-menu-item-modal">
+  <modal :align="'flex justify-end'" :show="showModal" class="add-new-menu-item-modal">
     <div slot="container">
       <div class="flex flex-wrap justify-between mb-6">
         <h2 class="text-90 font-normal text-xl">{{ __('Add Menu item') }}</h2>
-        <toggle-button v-model="newItem.enabled" :color="switchColor" :width="70" :sync="true" :labels="toggleLabels" />
+        <toggle-button :color="switchColor" :labels="toggleLabels" :sync="true" :width="70" v-model="newItem.enabled" />
       </div>
 
-      <form autocomplete="off" @submit.prevent="$emit(update ? 'updateItem' : 'confirmItemCreate')">
+      <form @submit.prevent="$emit(update ? 'updateItem' : 'confirmItemCreate')" autocomplete="off">
         <div class="flex border-b border-40">
           <div class="w-1/5 py-4">
             <label class="inline-block text-80 pt-2 leading-tight">{{ __('Name') }}</label>
           </div>
           <div class="py-4 w-4/5">
             <input
-              v-model="newItem.name"
-              id="name"
-              type="text"
               :placeholder="this.__('Name')"
               class="w-full form-control form-input form-input-bordered"
+              id="name"
+              type="text"
+              v-model="newItem.name"
             />
           </div>
         </div>
@@ -31,12 +31,13 @@
               @input="e => $emit('onLinkTypeUpdate', e.target.value)"
               class="w-full form-control form-select"
             >
-              <option value="" selected="selected" disabled="disabled">{{ __('Choose an option') }}</option>
+              <option disabled="disabled" selected="selected" value="">{{ __('Choose an option') }}</option>
 
-              <option :value="type.class" v-for="(type, i) of linkTypes" :key="i">{{ __(type.name) }}</option>
+              <option :key="i" :value="type.class" v-for="(type, i) of linkTypes">{{ __(type.name) }}</option>
             </select>
           </div>
         </div>
+
         <template v-if="linkType.type === 'static-url'">
           <div class="flex border-b border-40">
             <div class="w-1/5 py-4">
@@ -44,11 +45,11 @@
             </div>
             <div class="py-4 w-4/5">
               <input
-                v-model="newItem.value"
-                id="url"
-                type="text"
                 :placeholder="__('URL')"
                 class="w-full form-control form-input form-input-bordered"
+                id="url"
+                type="text"
+                v-model="newItem.value"
               />
             </div>
           </div>
@@ -62,21 +63,35 @@
 
             <div class="py-4 w-4/5">
               <multiselect
-                track-by="id"
-                label="label"
-                :value="options.find(option => option.id === newItem.value)"
                 :options="options"
+                :value="options.find(option => option.id === newItem.value)"
                 @input="value => $emit('onLinkModelUpdate', value.id)"
+                label="label"
+                track-by="id"
               />
             </div>
           </div>
         </template>
+
+        <card v-if="linkType.fields">
+          <component
+            :class="{ 'remove-bottom-border': index === linkType.fields.length - 1 }"
+            :field="field"
+            :is="`form-${field.component}`"
+            :key="index"
+            :resource-id="resourceId"
+            :resource-name="resourceName"
+            class="menu-item-component"
+            v-for="(field, index) in fields"
+          />
+        </card>
+
         <div class="flex border-b border-40">
           <div class="w-1/5 py-4">
             <label class="inline-block text-80 pt-2 leading-tight">{{ __('Parameters') }}</label>
           </div>
           <div class="py-4 w-4/5">
-            <codemirror v-model="newItem.parameters" :options="cmOptions" :placeholder="cmPlaceholder"></codemirror>
+            <codemirror :options="cmOptions" :placeholder="cmPlaceholder" v-model="newItem.parameters"></codemirror>
           </div>
         </div>
         <div class="flex border-b border-40" v-if="linkType.type && linkType.type !== 'text'">
@@ -84,7 +99,7 @@
             <label class="inline-block text-80 pt-2 leading-tight">{{ __('Open in') }}</label>
           </div>
           <div class="py-4 w-4/5">
-            <select v-model="newItem.target" class="w-full form-control form-select">
+            <select class="w-full form-control form-select" v-model="newItem.target">
               <option value="_self">{{ __('Same window') }}</option>
               <option value="_blank">{{ __('New window') }}</option>
             </select>
@@ -95,18 +110,22 @@
     <div slot="buttons">
       <div class="ml-auto">
         <button
-          type="button"
           @click.prevent="$emit('closeModal')"
           class="btn text-80 font-normal h-9 px-3 mr-3 btn-link"
+          type="button"
         >
           {{ __('Cancel') }}
         </button>
 
-        <button v-if="update === false" @click.prevent="$emit('confirmItemCreate')" class="btn btn-default btn-primary">
+        <button
+          @click.prevent="storeWithData('confirmItemCreate')"
+          class="btn btn-default btn-primary"
+          v-if="update === false"
+        >
           {{ __('Create menu item') }}
         </button>
 
-        <button v-else @click.prevent="$emit('updateItem')" class="btn btn-default btn-primary">
+        <button @click.prevent="storeWithData('updateItem')" class="btn btn-default btn-primary" v-else>
           {{ __('Update menu item') }}
         </button>
       </div>
@@ -124,7 +143,7 @@ import 'codemirror/theme/dracula.css';
 import 'codemirror/mode/javascript/javascript';
 
 export default {
-  props: ['newItem', 'showModal', 'update', 'linkType', 'linkTypes'],
+  props: ['newItem', 'showModal', 'update', 'linkType', 'linkTypes', 'resourceName', 'resourceId'],
   components: {
     Modal,
     codemirror,
@@ -147,15 +166,33 @@ export default {
     },
     cmPlaceholder: '{\n  "exampleValue": 5\n}',
   }),
+
   mounted() {
     this.toggleLabels = { checked: this.__('Enabled'), unchecked: this.__('Disabled') };
     this.switchColor = { checked: '#21b978', unchecked: '#dae1e7', disabled: '#eef1f4' };
   },
+
   computed: {
+    fields() {
+      return this.newItem.fields || this.linkType.fields;
+    },
+
     options() {
       const options = Object.keys(this.linkType.options).map(id => ({ id, label: this.linkType.options[id] }));
       options.unshift({ id: '', label: this.__('Choose an option') });
       return options;
+    },
+  },
+
+  methods: {
+    storeWithData(eventType) {
+      this.fields.forEach(field => {
+        const formData = new FormData();
+        field.fill(formData);
+        this.newItem[field.attribute] = formData.get(field.attribute);
+      });
+
+      this.$emit(eventType);
     },
   },
 };
@@ -170,6 +207,18 @@ export default {
 
     .CodeMirror-placeholder {
       color: rgba(#6272a4, 0.7);
+    }
+  }
+
+  .menu-item-component {
+    div:nth-child(1) {
+      padding-left: 0;
+    }
+
+    div:nth-child(2) {
+      padding-left: 0;
+      padding-right: 0;
+      width: 80%;
     }
   }
 }