+++ src/Classes/MenuItemStaticURL.php
@@ -2,6 +2,10 @@
 
 namespace OptimistDigital\MenuBuilder\Classes;
 
+use Illuminate\Http\Request;
+use Laravel\Nova\Fields\Text;
+use Laravel\Nova\Panel;
+
 class MenuItemStaticURL extends MenuLinkable
 {
     public static function getIdentifier(): string
@@ -17,5 +21,12 @@ class MenuItemStaticURL extends MenuLinkable
     public static function getType(): string
     {
         return 'static-url';
+    }
+
+    public static function getRules(): array
+    {
+        return [
+            'value' => 'required',
+        ];
     }
 }