+++ CHANGELOG.md
@@ -5,6 +5,25 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.0.0] - 2020-08-21
+
+Major release due to breaking changes!
+
+### Added
+
+- Ability to add custom fields to MenuLinkable class (via `getFields(): array`) function
+- Ability to customize rules for menu items (via `getRules(): array`) function
+- Added new `data` column to menu items table (`php artisan migrate` required)
+- Added a fourth parameter `$data` to `getDisplayValue()`
+
+```
+public static function getDisplayValue($value = null, array $parameters = null, array $data = null)
+```
+
+### Changed
+
+- Updated packages
+
 ## [2.3.8] - 2020-06-29
 
 ### Changed